requirements.txt
blinker==1.9.0
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.0
greenlet==3.2.2
holidays==0.73
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mysql-connector-python==9.3.0
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.1
prophet==1.1.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.41
stanio==0.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
